Function blind_bid::score_gen::score::compute_score_gadget

pub fn compute_score_gadget(
    composer: &mut StandardComposer, 
    bid: &Bid, 
    bid_value: Variable, 
    secret_k: Variable, 
    bid_tree_root: Variable, 
    consensus_round_seed: Variable, 
    latest_consensus_round: Variable, 
    latest_consensus_step: Variable
) -> Variable

Computes the score of the bid printing in the ConstraintSystem the proof of the correct obtention of the score.

Takes 3615 constraints.